добавляет фотографии в карточку
