+++ js/main.js
@@ -1,72 +1,5 @@
-import { getObjects, createObject } from './data.js';
+import { getObjects} from './data.js';
+import { createCardElement } from './card.js';
 
 getObjects();
-
-const typesHome = {
-  flat: 'Квартира',
-  bungalow: 'Бунгало',
-  house: 'Дом',
-  palace: 'Дворец',
-  hotel: 'Отель',
-};
-
-const renderFeatures = (cardElement, features) => {
-  const featuresList = cardElement.querySelector('.popup__features');
-  const featuresItem = featuresList.querySelectorAll('.popup__feature');
-  if (features && features.length) {
-    const modifiers = features.map((feature) => `popup__feature--${feature}`);
-    featuresItem.forEach((featuresListItem) => {
-      const modifier = featuresListItem.classList[1];
-      if (!modifiers.includes(modifier)) {
-        featuresListItem.remove();
-      }
-    });
-  } else {
-    featuresList.remove();
-  }
-};
-
-const renderDescription = (cardElement, description) => {
-  const descriptionElement = cardElement.querySelector('.popup__description');
-  if (description && description.length) {
-    descriptionElement.textContent = description;
-  } else {
-    descriptionElement.remove();
-  }
-};
-
-const cardTemplate = document.querySelector('#card')
-  .content
-  .querySelector('.popup');
-
-const createCardElement = ({ offer }) => {
-  const cardElement = cardTemplate.cloneNode(true);
-  cardElement.querySelector('.popup__title').textContent = offer.title;
-  cardElement.querySelector('.popup__text--address').textContent = offer.addres;
-  cardElement.querySelector('.popup__text--price').textContent = `${offer.price} ₽/ночь`;
-  cardElement.querySelector('.popup__type').textContent = typesHome[offer.type];
-  cardElement.querySelector('.popup__text--capacity').textContent = `${offer.rooms} комнаты для ${offer.guests} гостей»`;
-  cardElement.querySelector('.popup__text--time').textContent = `Заезд после ${offer.checkin}, выезд до ${offer.checkout}`;
-
-
-  renderDescription(cardElement, offer.description);
-  renderFeatures(cardElement, offer.features);
-
-  const mapContainer = document.querySelector('#map-canvas');
-  mapContainer.appendChild(cardElement);
-};
-
-createCardElement(
-  createObject(1)
-);
-
-
-// console.log(createCardElement(cardData));
-
-
-// На основе временных данных для разработки и шаблона #card создайте DOM-элементы, соответствующие объявлениям, и заполните их данными:
-
-// В блок .popup__photos выведите все фотографии из списка offer.photos. Каждая из строк массива photos должна записываться как атрибут src соответствующего изображения.
-
-// Замените значение атрибута src у аватарки пользователя .popup__avatar на значение поля author.avatar.
-
+createCardElement();
